requirements.txt
certifi==2022.12.7
charset-normalizer==3.0.1
greenlet==2.0.1
idna==3.4
psycopg2-binary==2.9.5
pyTelegramBotAPI==4.11.0
python-dotenv==0.21.1
requests==2.30.0
SQLAlchemy==2.0.13
typing_extensions==4.4.0
urllib3==1.26.14
